add crm association table extension
