+++ crm-association-table-example/hsproject.json
@@ -1,0 +1,4 @@
+{
+  "name": "example-crm-associations-table",
+  "srcDir": "src"
+}
