+++ crm-association-table-example/src/app/app.functions/package.json
