+++ crm-association-table-example/src/app/app.json
@@ -1,5 +1,5 @@
 {
-  "name": "CRM Components React App",
+  "name": "CRM Association Table React App",
   "description": "This is an example of private app that shows a custom card. It demonstrates usages of CRM Components",
   "scopes": ["crm.objects.contacts.read", "crm.objects.contacts.write"],
   "public": false,
@@ -7,7 +7,7 @@
     "crm": {
       "cards": [
         {
-          "file": "extensions/example-card.json"
+          "file": "extensions/association-table.json"
         }
       ]
     }
