+++ crm-association-table-example/src/app/extensions/AssociationTableExtension.jsx
@@ -1,0 +1,55 @@
+import React from 'react';
+import {
+  hubspot,
+  Text,
+} from '@hubspot/ui-extensions';
+import { CrmAssociationTable } from '@hubspot/ui-extensions/crm';
+
+// Define the extension to be run within the Hubspot CRM
+hubspot.extend(({ context, runServerlessFunction, actions }) => (
+  <Extension
+    context={context}
+    runServerless={runServerlessFunction}
+    sendAlert={actions.addAlert}
+  />
+));
+
+const Extension = ({ context, runServerless, sendAlert }) => {
+  const run = () => {
+    runServerless({ name: 'myFunc', parameters: { text: text } }).then((resp) =>
+      sendAlert({ message: resp.response })
+    );
+  };
+
+  return (
+    <>
+       <Text format={{ fontWeight: 'bold' }}>Important Deals</Text>
+       <CrmAssociationTable
+        objectTypeId="0-3"
+        propertyColumns={["dealname", "hubspot_owner_id", "amount", "dealstage"]}
+        quickFilterProperties={["hubspot_owner_id", "dealstage", "amount"]}
+        pageSize={10}
+        preFilters={[
+          {
+            operator: "NOT_IN",
+            property: "dealstage",
+            values: ["closedwon", "closedlost"]
+          },
+          {
+            operator: "GT",
+            property: "amount",
+            value: 1000
+          }
+        ]}
+        sort={[
+          {
+            direction: 1,
+            columnName: 'amount',
+          },
+        ]}
+        searchable={true}
+        pagination={true}
+      />
+    </>
+  );
+};
