--- crm-components-example/hsproject.json
@@ -1,4 +1,0 @@
-{
-  "name": "CRM Components React App",
-  "srcDir": "src"
-}
